+++ src/cgeniepy/skill.py
@@ -2,11 +2,17 @@ import numpy as np
 from scipy.spatial import distance
 import matplotlib.pyplot as plt
 
+from matplotlib.projections import PolarAxes
+import mpl_toolkits.axisartist.floating_axes as fa
+import mpl_toolkits.axisartist.grid_finder as gf
+
+import warnings
+
 class ArrComparison:
     
     "quantitatively compare similarity metrics between two 2D arrays (model and observation)"
 
-    def __init__(self, model, observation):
+    def __init__(self, model, observation, label=None):
         """
         
         :param: model: array-like data
@@ -29,6 +35,7 @@ class ArrComparison:
         """
         self.model = model
         self.data = observation
+        self.label = label
 
         ## check data type, if not float, convert to float
         if self.model.dtype != np.float64:
@@ -176,171 +183,134 @@ class DFComparison(ArrComparison):
         super().__init__(self.model, self.data)
 
  
-# class TaylorDiagram(object):
-#     """
-#     Taylor diagram.
-#     Plot model standard deviation and correlation to reference data in a single-quadrant polar plot,
-#     with r=stddev and theta=arccos(correlation).
+class TaylorDiagram(object):
+    """
+    Taylor diagram.
 
-#     modified from Yannick Copin, https://gist.github.com/ycopin/3342888
-#     reference: https://matplotlib.org/stable/gallery/axisartist/demo_floating_axes.html
-#     """
+    A visualisation of model-data comparison considering (1) RMSE; (2) Correlation; (3) Standard Deviation.
 
-#     def __init__(
-#             self,
-#             fig=None,
-#             figscale=1,
-#             subplot=111,
-#             xmax=None,
-#             tmax=np.pi / 2,
-#             ylabel="Standard Deviation",
-#             rotation=None,
-#     ):
+    In practical it is a polar axis with point distributed at (std, arccos(correlation))
+    These two metrics, combining with the observed standard deviation, drived the cRMSE that measures the distance from reference point
 
-#         """
-#         Set up Taylor diagram axes, i.e. single quadrant polar
-#         plot, using `mpl_toolkits.axisartist.floating_axes`.
+    see https://en.wikipedia.org/wiki/Taylor_diagram#/media/File:Taylor_diagram_fig2.png
+    and https://bookdown.org/david_carslaw/openair/sections/model-evaluation/taylor-diagram.html (Figure 20.2)
 
-#         Parameters:
+    This class is modified from Yannick Copin, https://gist.github.com/ycopin/3342888
+    """
 
-#         * fig: input Figure or None
-#         * subplot: subplot definition
-#         * xmax: the length of radius, xmax can be 1.5* reference std
-#         """
+    def __init__(self, ac: ArrComparison):
+        if isinstance(ac, list):
+            self.mult_comp = True
+        else:
+            self.mult_comp = False
+            
+        self.ac = ac
+        self._extract_data()
 
-#         # --------------- tickers --------------------------
-#         # Correlation labels (if half round)
-#         cor_label = np.array([0, 0.2, 0.4, 0.6, 0.8, 0.9, 0.95, 0.99, 1])
+    def _extract_data(self):
 
-#         # add the negative ticks if more than half round
-#         excess_theta = tmax - np.pi / 2
-#         if excess_theta > 0:
-#             cor_label = np.concatenate((-cor_label[:0:-1], cor_label))
+        if not self.mult_comp:        
+            self.corr = self.ac.pearson_r()
+            self.model_std = np.std(self.ac.model)
[…]
