### 2024-04-24 da3690d
more taylordiagram optimisation
