+++ docs/source/visualisation.rst
@@ -31,5 +31,5 @@ I have to admit it is not very smart to do the customisation. Instead of like `m
 
 
 More Examples
------------
+------------------
 Please see the gallery section.
