+++ examples/plot_PO4_distribution.py
@@ -3,23 +3,35 @@
 Plot 2D transect of tracers in each basin
 =========================================
 
-This example shows how to plot the PO4 distribution in each basin.
+This example plots the modelled PO4 distribution in cGENIE.
+
+The following features in the package are used:
+* Access data through `cgeniepy.model` module
+* A basin-mask operation
+* A linear interpolation
+* Additional color palette (mirrors the one in ODV)
+* Self-optimised plotting method
 """
 
 from cgeniepy.model import GenieModel
+from cgeniepy.plot import CommunityPalette
 import matplotlib.pyplot as plt
 
 model = GenieModel("/Users/yingrui/Science/lgm_foram_niche/model/muffin.CBE.worlg4.BASESFeTDTL.SPIN")
 ocn_po4 = model.get_var("ocn_PO4").isel(time=-1)
 
-fig, axs=plt.subplots(nrows=1, ncols=3, figsize=(15, 3), tight_layout=True)        
+fig, axs=plt.subplots(nrows=1, ncols=3, figsize=(15, 3), tight_layout=True)
 
 basins = ['Atlantic', 'Pacific', 'Indian']
 
-for i in range(3):
-	basin_data = model.get_var('ocn_PO4').isel(time=-1).mask_basin(base='worjh2',basin=basins[i], subbasin='')
-	basin_data.data.values = basin_data.data.values * 1E6
-	basin_data.mean(dim='lon').interpolate().plot(ax=axs[i], contour=True)
-	axs[i].title.set_text(basins[i])
+odv_cmap = CommunityPalette().get_palette('ODV', reverse=True)
 
-plt.show()        
+for i in range(3):
+    basin_data = model.get_var('ocn_PO4').isel(time=-1).mask_basin(base='worjh2',basin=basins[i], subbasin='')
+    basin_data.data.values = basin_data.data.values * 1E6
+    basin_data_interp = basin_data.mean(dim='lon').interpolate().to_GriddedDataVis()
+    basin_data_interp.aes_dict['pcolormesh_kwargs']['cmap'] = odv_cmap
+    basin_data_interp.plot(ax=axs[i], contour=True)
+    axs[i].title.set_text(basins[i])
+
+plt.show()
