### 2024-04-24 fc3603c
rebuild plot_PO4_distribution example
