+++ setup.cfg
@@ -32,4 +32,4 @@ install_requires =
 where = src
 
 [options.package_data]
-* = *.nc, *.csv, *.txt
+* = *.nc, *.csv, *.txt, *.xml
