+++ docs/source/auto_examples/plot_PO4_distribution.rst
@@ -36,7 +36,7 @@ The following features in the package are used:
 
 #. Customise the plotting details
 
-.. GENERATED FROM PYTHON SOURCE LINES 20-43
+.. GENERATED FROM PYTHON SOURCE LINES 20-42
 
 
 
@@ -79,7 +79,6 @@ The following features in the package are used:
 
     for i in range(3):
         basin_data = model.get_var('ocn_PO4').isel(time=-1).mask_basin(base='worjh2',basin=basins[i], subbasin='')
-        basin_data.data.values = basin_data.data.values * 1E6
         basin_data_interp = basin_data.mean(dim='lon').interpolate().to_GriddedDataVis()
         basin_data_interp.aes_dict['pcolormesh_kwargs']['cmap'] = odv_cmap
         basin_data_interp.plot(ax=axs[i], contour=True)
@@ -90,7 +89,7 @@ The following features in the package are used:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 1.789 seconds)
+   **Total running time of the script:** (0 minutes 1.661 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_PO4_distribution.py:
