### 2024-04-25 0ca95d1
update the plot_multiple_vars.py
