+++ examples/plot_multiple_vars.py
@@ -4,6 +4,14 @@ Plot multiple variables
 =======================================================
 
 This example shows how to plot multiple variables from a GenieModel object. I use a modern model run as an example and plot the surface temperature, PO4, Fe, and O2.
+
+The following features are particularly demonstrated:
+
+#. Customizing the plot projection with cartopy
+
+#. Plotting multiple variables in one figure
+
+#. Controlling the elements in cgeniepy.GriddedDataVis object
 """
 
 ## many variables
@@ -18,4 +26,6 @@ pi_model = GenieModel("/Users/yingrui/Science/lgm_foram_niche/model/muffin.CBE.w
 variable = ['ocn_sur_temp', 'ocn_sur_PO4', 'ocn_sur_TDFe', 'ocn_sur_O2']
 
 for i in range(4):
-    pi_model.get_var(variable[i]).isel(time=-1).plot(ax=axs.flatten()[i],colorbar=True, outline=True, contourf=True)
+    pi_model.get_var(variable[i]).isel(time=-1).plot(ax=axs.flatten()[i],colorbar=True,
+                                                     outline=True, contourf=True,
+                                                     gridline=True)
