### 2024-04-28 e0993e6
correct the model label in example
