### 2024-04-28 19e98f4
update example rendering outputs
