### 2024-04-28 2c908d7
save the attribute in add/substract
