### 2024-04-28 8876436
fix bug: check attribution using np.test.assrt_equal
