### 2024-05-01 7c8e735
add calculation method in array.py
