### 2024-05-03 0744da8
add get_genie_lon/lat/zt back
