### 2024-05-03 057c84d
optimise lon_n2g: add offset option
