### 2024-05-03 6c3b713
simpler to_geniebin method in ScatterData class
