+++ src/cgeniepy/table.py
@@ -164,88 +164,34 @@ class ScatterData:
     def to_geniebin(
         self,
         var,
-        low_threshold=None,
-        new_low_bound=None,
-        high_threshold=None,
-        new_high_bound=None,
     ):
         """
         Regrid a dataframe within certain format to cGENIE grids
 
-        Input: A dataframe with three necessary columns: "Latitude", "Longitude", "Observation".
-        The other columns will be unselected. Note that observation column is the data you want to process.
-        Rename dataframe by using command: `df = df.rename({"old name": "new name"}, axis='columns')`
-
         Output:
         A 36x36 2D array.
-
-        Optional parameters:
-        low_threshold/new_low_bound: change the data that is smaller than low_threshold to new_low_bound.
-        For example, covert any data < 0.03 to 0. Work same to high_threshold/new_high_bound.
         """
-
-        assert self.lon in self.index, f"{self.lon} not found in the index"
-        assert self.lat in self.index, f"{self.lat} not found in the index"
 
         go= GridOperation()
         
-        df = self.data.copy()
+        src_df = self.data.reset_index(inplace=False)
 
         # subset
-        df = df[[self.lon, self.lat, var]]
+        src_df = src_df[[self.lon, self.lat, var]]
 
         # drop NAN
-        df = df.dropna(axis="rows", how="any")
+        src_df = src_df.dropna(axis="rows", how="any")
 
-        # regrid coordinate
-        df.loc[:, self.lon] = df.loc[:, self.lat].apply(go.geniebin_lon)
-        df.loc[:, self.lon] = df.loc[:, self.lon].apply(go.geniebin_lat)
+        # regrid coordinate: genie lat x normal lon
+        ## lat
+        src_df.loc[:, self.lat] = src_df.loc[:, self.lat].apply(go.geniebin_lat)
+        ## lon 
+        src_df.loc[:, self.lon] = src_df.loc[:, self.lon].apply(go.normbin_lon)
 
-        # group and aggregate (still in long format)
-        df_agg = df.groupby([self.lon, self.lat]).agg("mean")
+        # aggregated source data (in long format)
+        src_df_agg = src_df.groupby([self.lat,self.lon]).agg("mean")
+        return src_df_agg
 
-        # Create a all-nan data frame with full cGENIE grids
-        lat = go.get_genie_lat()
-        lon = go.get_normal_lon()
-        data = np.zeros([36 * 36])
-        index = pd.MultiIndex.from_product([lon, lat], names=[self.lon, self.lat])
-        df_genie = pd.DataFrame(data, index=index, columns=[var])
-        df_genie[var] = df_genie[var].replace(0, np.nan)
-
-        # transform the tuple elements in multindex to lists
-        df_genie_index_list = [list(item) for item in df_genie.index.values]
-        df_agg_index_list = [list(item) for item in df_agg.index.values]
-
-        # copy aggregated dataframe to full-grid dataframe, one by one
-        # if the returned dataframe is full of NA, it's very likely from this step
-        for i in df_genie_index_list:
-
-            longitude = i[0]
-            latitude = i[1]
-
-            if i in df_agg_index_list:
-                df_genie.loc[longitude, latitude] = df_agg.loc[longitude, latitude]
-            else:
-                df_genie.loc[longitude, latitude] = np.nan
-
-        # filter data if necessary
-        if (low_threshold is not None) and (new_low_bound is not None):
-            df_genie.loc[
-                df_genie.Observation < low_threshold, var
-            ] = new_low_bound
-
-        if (high_threshold is not None) and (new_high_bound is not None):
-            df_genie.loc[
-                df_genie.Observation > high_threshold, var
-            ] = new_high_bound
-
[…]
