+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = cgeniepy
-version = 0.12.4
+version = 0.13.0
 author = Rui Ying
 author_email = rui.ying@bristol.ac.uk
 description = A Python package to read, analyse and visualise cGENIE Earth System Model output
