### 2024-05-03 6814c10
make to_geniebin available for 3D data
