### 2024-05-03 205b0a8
remove hard code in get_genie_lat
