### 2024-05-03 bcc53da
add __array__ method
