### 2024-05-04 9b3a291
use hashtable to replace if else
