### 2024-05-04 d2af49c
optimise the get_diag_avg method
