### 2024-05-04 0650c87
support index year of diagnostic files
