### 2024-05-04 bee30e6
support model_ensemble in get_diag_avg method
