### 2024-05-04 5e518a3
add __repr__ for GenieModel
