### 2024-05-04 a2a8561
add __repr__ for GriddedData
