### 2024-05-04 deb2e99
add __repr__ for ScatterData
