### 2024-05-04 f3f7092
add a new example of plot_regridded_data.py
