+++ docs/source/auto_examples/sg_execution_times.rst
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:01.661** total execution time for 14 files **from auto_examples**:
+**00:02.809** total execution time for 15 files **from auto_examples**:
 
 .. container::
 
@@ -32,8 +32,11 @@ Computation times
    * - Example
      - Time
      - Mem (MB)
+   * - :ref:`sphx_glr_auto_examples_plot_regridded_data.py` (``plot_regridded_data.py``)
+     - 00:02.809
+     - 0.0
    * - :ref:`sphx_glr_auto_examples_plot_PO4_distribution.py` (``plot_PO4_distribution.py``)
-     - 00:01.661
+     - 00:00.000
      - 0.0
    * - :ref:`sphx_glr_auto_examples_plot_additional_cmap.py` (``plot_additional_cmap.py``)
      - 00:00.000
