+++ docs/source/auto_examples/plot_regridded_data.rst
@@ -1,0 +1,130 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/plot_regridded_data.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_plot_regridded_data.py>`
+        to download the full example code
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_plot_regridded_data.py:
+
+
+==========================================================
+Regrid observational data and compare with cGENIE output
+==========================================================
+
+This example shows how to regrid observational data to cGENIE grids and compare with cGENIE output.
+
+This example uses many features from the cgeniepy package, including:
+
+#. GirdOperation (eastern to normal longitude conversion)
+
+#. Read cGENIE output
+
+#. Bin observational data to cGENIE grid (a coarse implementation)
+
+#. Compare model and observational data
+
+.. GENERATED FROM PYTHON SOURCE LINES 18-57
+
+
+
+.. image-sg:: /auto_examples/images/sphx_glr_plot_regridded_data_001.png
+   :alt: plot regridded data
+   :srcset: /auto_examples/images/sphx_glr_plot_regridded_data_001.png
+   :class: sphx-glr-single-img
+
+
+.. rst-class:: sphx-glr-script-out
+
+ .. code-block:: none
+
+    <frozen importlib._bootstrap>:241: RuntimeWarning: scipy._lib.messagestream.MessageStream size changed, may indicate binary incompatibility. Expected 56 from C header, got 64 from PyObject
+    /Users/yingrui/cgeniepy/src/cgeniepy/model.py:50: UserWarning: No gemflag is provided, use default gemflags: [biogem]
+      warnings.warn("No gemflag is provided, use default gemflags: [biogem]")
+    M-score of sea surface temperature 0.8211028102365528
+
+
+
+
+
+
+|
+
+.. code-block:: Python
+
+    import xarray as xr
+    import numpy as np
+    import matplotlib.pyplot as plt
+    import cartopy.crs as ccrs
+
+    from cgeniepy.table import ScatterData
+    from cgeniepy.grid import GridOperation
+    from cgeniepy.array import GriddedData
+    from cgeniepy.model import GenieModel
+    from cgeniepy.skill import ArrComparison
+
+    ## load GLODAP temperature data
+    glodap_temp = xr.load_dataset("/Users/yingrui/Science/GLODAPv2.2016b_MappedClimatologies/GLODAPv2.2016b.temperature.nc")['temperature']
+    ## convert to normal longitude from eastern longitude
+    glodap_temp = GridOperation().xr_e2n(glodap_temp)
+    ## convert to ScatterData
+    glodap_temp = ScatterData(glodap_temp.isel(depth_surface=0).to_dataframe())
+    ## bin into genie coordinate and convert back to xarray
+    glodap_temp = glodap_temp.to_geniebin(var='temperature').to_xarray()['temperature']
+
+    ## This is cGENIE output
+    model = GenieModel("/Users/yingrui/Science/lgm_foram_niche/model/muffin.CBE.worlg4.BASESFeTDTL.SPIN")
+    worlg4_sst = model.get_var('ocn_sur_temp').isel(time=-1).normalise_longitude(method='g2n')
+
[…]
