### 2024-05-04 a92f06f
add one more feature in plot_regridded_data.py
