### 2024-05-04 329bb8e
revise the syntax of plot_logo.py
