### 2024-05-06 c4bfaa5
fix a bug: importlib.resource
