### 2024-05-06 355bc8b
add MANIFEST.in to add external data
