+++ docs/source/conf.py
@@ -15,7 +15,7 @@ sys.path.insert(0, os.path.abspath('../src/cgeniepy/'))
 project = 'cgeniepy'
 copyright = f"2024-{datetime.datetime.now().year}, Rui Ying"
 author = 'Rui Ying'
-release = '0.13.0'
+release = '0.13.1'
 
 
 # -- General configuration ---------------------------------------------------
