+++ docs/source/auto_examples/plot_scatterdata.py
@@ -7,6 +7,6 @@ This example shows how to plot the 1D ScatterData object. I use a data file from
 """
 
 from cgeniepy.table import ScatterData
-data= ScatterData("/Users/yingrui/cgeniepy/test/EDC_CO2.tab", sep='\t')
+data= ScatterData("/Users/yingrui/cgeniepy/src/data/EDC_CO2.tab", sep='\t')
 data.set_index(['Age [ka BP]'])
 data.to_ScatterDataVis().plot(var='CO2 [ppmv]')
