+++ docs/source/auto_examples/plot_other_models.rst
@@ -39,7 +39,7 @@ This example shows how to use cgeniepy to plot gridded data from other models.
  .. code-block:: none
 
 
-    <matplotlib.colorbar.Colorbar object at 0x17cdc7d10>
+    <matplotlib.colorbar.Colorbar object at 0x3449d3a30>
 
 
 
@@ -77,7 +77,7 @@ This example shows how to use cgeniepy to plot gridded data from other models.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 5.071 seconds)
+   **Total running time of the script:** (0 minutes 4.566 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_other_models.py:
