+++ docs/source/auto_examples/plot_d13C_comp.rst
@@ -81,7 +81,7 @@ This example is a combination of ScatterData and GriddedData to search the neare
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 33.478 seconds)
+   **Total running time of the script:** (0 minutes 30.557 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_d13C_comp.py:
