### 2024-05-06 b83ed19
add 1D interpolation method
