### 2024-05-06 41a435b
add rolling mean method and related example
