### 2024-05-06 5c2e52f
use spectral_r as cmap in sst2Dmap example
