+++ docs/source/auto_examples/plot_multiple_vars.rst
@@ -67,7 +67,7 @@ The following features are particularly demonstrated:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 1.233 seconds)
+   **Total running time of the script:** (0 minutes 1.262 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_multiple_vars.py:
