+++ docs/source/auto_examples/plot_arrcomparison.rst
@@ -38,6 +38,7 @@ This example shows how to use the ArrComparison class to compare two arrays and 
 
  .. code-block:: none
 
+    <frozen importlib._bootstrap>:241: RuntimeWarning: scipy._lib.messagestream.MessageStream size changed, may indicate binary incompatibility. Expected 56 from C header, got 64 from PyObject
     Pearson coefficient: 0.7054519438385228
 
 
@@ -65,7 +66,7 @@ This example shows how to use the ArrComparison class to compare two arrays and 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.147 seconds)
+   **Total running time of the script:** (0 minutes 0.213 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_arrcomparison.py:
