+++ docs/source/auto_examples/plot_logo.rst
@@ -72,7 +72,7 @@ This example shows how to customise the 2D map including the projection, the col
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.228 seconds)
+   **Total running time of the script:** (0 minutes 0.249 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_logo.py:
