+++ docs/source/auto_examples/plot_additional_cmap.rst
@@ -78,7 +78,7 @@ ODV's palette.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.212 seconds)
+   **Total running time of the script:** (0 minutes 0.205 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_additional_cmap.py:
