+++ docs/source/auto_examples/plot_PO4_distribution.rst
@@ -88,7 +88,7 @@ The following features in the package are used:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.341 seconds)
+   **Total running time of the script:** (0 minutes 0.347 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_PO4_distribution.py:
