+++ docs/source/auto_examples/plot_scatterdata.rst
@@ -34,18 +34,8 @@ This example shows how to plot the 1D ScatterData object. I use a CO2 data file 
    :class: sphx-glr-single-img
 
 
-.. rst-class:: sphx-glr-script-out
-
- .. code-block:: none
-
-    <frozen importlib._bootstrap>:241: RuntimeWarning: scipy._lib.messagestream.MessageStream size changed, may indicate binary incompatibility. Expected 56 from C header, got 64 from PyObject
 
 
-
-
-
-
-|
 
 .. code-block:: Python
 
@@ -80,7 +70,7 @@ This example shows how to plot the 1D ScatterData object. I use a CO2 data file 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.886 seconds)
+   **Total running time of the script:** (0 minutes 0.067 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_scatterdata.py:
