+++ docs/source/auto_examples/plot_taylor_diagram.rst
@@ -82,7 +82,7 @@ The R package open-air has provided a nice explanation on what taylor diagram is
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.070 seconds)
+   **Total running time of the script:** (0 minutes 0.074 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_taylor_diagram.py:

[…]
