### 2024-05-06 6e4e34e
add minor ticks for transect plot
