+++ examples/plot_timeseries.py
@@ -1,19 +1,25 @@
 """
-================
+====================
 Plot Time Series
-================
+====================
 
 This example shows how to read in and plot time series data
 """
 from cgeniepy.model import GenieModel
+import seaborn as sns
 import matplotlib.pyplot as plt
 
 model = GenieModel("/Users/yingrui/Science/lgm_foram_niche/model/muffin.CBE.worlg4.BASESFeTDTL.historical")
-fig, ax = plt.subplots(1,2,figsize=(10,5))
+fig, axs = plt.subplots(1,2,figsize=(8, 3), tight_layout=True)
 
-model.get_ts("ocn_temp").plot(x='time (yr)',y='temperature (C)',ax=ax[0])
-model.get_ts("ocn_O2").plot(x='time (yr)',y='global total O2 (mol)',ax=ax[1])
+temp = model.get_ts("ocn_temp")
+o2 = model.get_ts("ocn_O2")
 
-ax[0].set_title("Ocean mean temperature")
-ax[1].set_title("Global total ocean oxygen")
-plt.show()
+## merge both
+merged = temp.merge(o2, on="time (yr)")
+
+## use seaborn to plot
+sns.set_theme(context='notebook', style='ticks', palette='deep')
+sns.lineplot(data=merged, x="time (yr)", y="temperature (C)", ax=axs[0])
+sns.lineplot(data=merged, x="time (yr)", y="global mean O2 (mol kg-1)", ax=axs[1])
+sns.despine()
