### 2024-05-07 48ff1c2
add back outline for transect plot
