### 2024-05-07 e477837
add zt_moc and lat_moc for genie dim condidates
