### 2024-05-07 5a3adbc
fix a bug where alpha/N not working in colormap
