### 2024-05-07 d12ca23
adjust the cmd order of returning cmap
