### 2024-05-07 47e6adb
remove testpypi in installation instruction
