### 2024-05-07 30c2604
remove some ugly colormaps
