+++ src/data/colormaps/WtBuGnYlRd.txt
@@ -1,0 +1,256 @@
+#FFFFFF
+#FCFEFF
+#FAFDFF
+#F7FCFE
+#F4FBFE
+#F2FAFE
+#EFF9FE
+#ECF8FD
+#EAF7FD
+#E7F6FD
+#E5F5FD
+#E2F4FD
+#DFF3FC
+#DDF2FC
+#DAF1FC
+#D7F0FC
+#D5EFFC
+#D2EEFB
+#CFEDFB
+#CDECFB
+#CAEBFB
+#C7EAFA
+#C5E9FA
+#C2E8FA
+#BFE7FA
+#BDE6FA
+#BAE5F9
+#B7E4F9
+#B5E3F9
+#B2E2F9
+#B0E1F9
+#ADE0F8
+#AADFF8
+#A8DEF8
+#A5DDF8
+#A2DCF7
+#9EDAF7
+#9BD8F6
+#98D6F5
+#96D4F3
+#94D2F2
+#92D0F1
+#8FCEF0
+#8DCCEE
+#8BCAED
+#89C8EC
+#86C6EB
+#84C3EA
+#82C1E8
+#7FBFE7
+#7DBDE6
+#7BBBE5
+#79B9E4
+#76B7E2
+#74B5E1
+#72B3E0
+#6FB1DF
+#6DAFDD
+#6BADDC
+#68ABDB
+#66A9DA
+#64A7D9
+#62A5D7
+#5FA3D6
+#5EA1D5
+#5B9FD4
+#599DD3
+#579BD2
+#5499D0
+#5297CF
+#5095CE
+#4E93CD
+#4A90CB
+#488EC8
+#488FC4
+#4890C1
+#4891BD
+#4892BA
+#4893B6
+#4894B3
+#4895AF
+#4896AB
+#4897A8
+#4899A4
+#489AA1
+#489B9E
+#489C9A
+#489D96
+#489E93
+#489F8F
+#48A08C
+#48A188
+#49A285
+#49A381
+#49A47E
+#49A57A
+#49A677
+#49A773
+#49A870
+#49A96C
+#49AA69
+#49AC65
+#49AD62
+#49AE5E
+#49AF5A
+#49B057
+#49B153
+#49B250
+#49B34B
+#49B546
+#4EB647
+#53B847
+#57B948
+#5CBB48
+#61BC49
+#66BD4A
+#6ABF4A
+#6FC04B
+#74C14B
+#79C34C
+#7EC44D
+#82C64D
+#87C74E
+#8CC84E
+#91CA4F
+#96CB50
+#9ACC50
+#9FCE51
+#A3CE51
+#A8D051
+#ACD152
+#B1D252
+#B6D453
+#BBD554
+#C0D654
+#C4D855
+#C9D955
+#CEDB56
+#D3DC57
+#D8DD57
+#DCDF58
+#E1E058
+#E6E159
+#EBE35A
+#EFE45A
+#F4E65B
+#F9E75B
+#FAE55B
+#FAE159
+#FADE58
+#F9DA56
+#F9D755
+#F9D454
+#F9D052
+#F9CD51
+#F9C950
+#F9C64E
+#F9C34D
+#F8BF4B
+#F8BC4A
+#F8B949
+#F8B547
+#F8B246
+#F8AF45
+#F8AB43
+#F7A842
+#F7A440
+#F7A13F
+#F79D3E
+#F79A3C
+#F7973B
+#F7943A
+#F69038
+#F68D37
+#F68A35
+#F68634
+#F68333
+#F67F31
+#F67C30
+#F6792F
+#F5752D
+#F5722C
+#F56C2A
+#F46829
+#F36629
+#F26429
+#F16229
+#F06029
+#EF5E29
+#EF5C29
+#EE5A29
+#ED5829
[…]
