+++ src/data/colormaps/w5m4.txt
@@ -1,0 +1,256 @@
+#28478A
+#28488B
+#294A8C
+#2A4C8D
+#2B4D8E
+#2C4F8F
+#2C5190
+#2D5292
+#2E5493
+#2F5694
+#305795
+#305996
+#315B97
+#325D99
+#335E9A
+#34609B
+#34629C
+#35639D
+#36659E
+#3767A0
+#3868A1
+#396AA2
+#396CA3
+#3A6DA4
+#3B6FA5
+#3C71A7
+#3D73A8
+#3D74A9
+#3E76AA
+#3F78AB
+#4179AC
+#427BAD
+#447CAE
+#457EAF
+#4780B0
+#4881B0
+#4A83B1
+#4B84B2
+#4C86B3
+#4E87B4
+#4F89B5
+#518BB6
+#528CB7
+#548EB7
+#558FB8
+#5791B9
+#5893BA
+#5A94BB
+#5B96BC
+#5C97BD
+#5E99BE
+#5F9BBF
+#619CBF
+#629EC0
+#649FC1
+#65A1C2
+#67A2C3
+#68A4C4
+#6AA5C4
+#6CA7C5
+#6EA8C5
+#70A9C6
+#72AAC7
+#74ACC7
+#76ADC8
+#78AEC8
+#7AB0C9
+#7CB1C9
+#7EB2CA
+#7FB4CA
+#81B5CB
+#83B6CC
+#85B8CC
+#87B9CD
+#89BACD
+#8BBBCE
+#8DBDCE
+#8FBECF
+#91BFD0
+#93C1D0
+#95C2D1
+#97C3D1
+#99C5D2
+#9BC6D2
+#9DC7D3
+#9FC9D4
+#A0C9D3
+#A2CAD2
+#A4CAD2
+#A6CBD1
+#A8CCD1
+#A9CCD0
+#ABCDCF
+#ADCECF
+#AFCECE
+#B1CFCE
+#B2CFCD
+#B4D0CC
+#B6D1CC
+#B8D1CB
+#BAD2CB
+#BBD3CA
+#BDD3C9
+#BFD4C9
+#C1D5C8
+#C3D5C8
+#C4D6C7
+#C6D6C6
+#C8D7C6
+#CAD8C5
+#CCD8C4
+#CDD9C4
+#CFDAC3
+#D1DAC3
+#D2DBC2
+#D3DBC1
+#D4DCC1
+#D5DCC0
+#D6DCBF
+#D8DDBE
+#D9DDBE
+#DADDBD
+#DBDEBC
+#DCDEBC
+#DDDFBB
+#DEDFBA
+#DFDFBA
+#E0E0B9
+#E1E0B8
+#E2E1B7
+#E3E1B7
+#E4E1B6
+#E5E2B5
+#E6E2B5
+#E7E3B4
+#E8E3B3
+#EAE3B3
+#EBE4B2
+#ECE4B1
+#EDE4B0
+#EEE5B0
+#EFE5AF
+#EFE5AE
+#EFE3AC
+#EEE1AB
+#EDDFA9
+#ECDDA7
+#ECDCA6
+#EBDAA4
+#EAD8A3
+#E9D6A1
+#E9D49F
+#E8D29E
+#E7D09C
+#E6CE9A
+#E6CD99
+#E5CB97
+#E4C996
+#E3C794
+#E3C592
+#E2C391
+#E1C18F
+#E0BF8D
+#E0BE8C
+#DFBC8A
+#DEBA89
+#DDB887
+#DDB685
+#DCB484
+#DBB282
+#DBB181
+#D9AE80
+#D8AC7F
+#D7AA7E
+#D6A87D
+#D4A67C
[…]

[…]
