### 2024-05-07 7ca9cdb
add repr for community palette
