### 2024-05-08 1cd07cd
not specify matplotlib version
this uses a new feature for testing, not necessary for functionality
