+++ setup.cfg
@@ -21,7 +21,7 @@ install_requires =
     xarray
     numpy
     pandas
-    matplotlib>=3.8.4
+    matplotlib
     cartopy
     netCDF4
     scipy
