### 2024-05-08 36f2657
support 3D plot now for quick view
