### 2024-05-08 f26cc02
fix a bug that reverse option is not properly working
