+++ docs/source/auto_examples/plot_multiple_vars.py
@@ -12,20 +12,27 @@ The following features are particularly demonstrated:
 #. Plotting multiple variables in one figure
 
 #. Controlling the elements in cgeniepy.GriddedDataVis object
+
+#. Using CommunityPalette to get a pretty color palette
 """
 
-## many variables
+
 import matplotlib.pyplot as plt
 import cartopy.crs as ccrs
 from cgeniepy.model import GenieModel
+from cgeniepy.plot import CommunityPalette
 
 fig, axs = plt.subplots(2, 2, figsize=(10, 8), subplot_kw={'projection': ccrs.PlateCarree()})
 
 pi_model = GenieModel("/Users/yingrui/Science/lgm_foram_niche/model/muffin.CBE.worlg4.BASESFeTDTL.SPIN", gemflag='biogem')
 
-variable = ['ocn_sur_temp', 'ocn_sur_PO4', 'ocn_sur_TDFe', 'ocn_sur_O2']
+variable = ['ocn_sur_temp', 'bio_export_POC', 'ocn_sur_PO4','ocn_sur_O2']
+cmap = ['ocean_temp', 'tol_rainbow', 'WtBuGnYlRd','kovesi_rainbow']
 
 for i in range(4):
-    pi_model.get_var(variable[i]).isel(time=-1).plot(ax=axs.flatten()[i],colorbar=True,
-                                                     outline=True, contourf=True,
-                                                     gridline=True)
+    data = pi_model.get_var(variable[i]).isel(time=-1).to_GriddedDataVis()
+    data.aes_dict['pcolormesh_kwargs']['cmap'] = CommunityPalette(cmap[i]).colormap
+    data.aes_dict['contourf_kwargs']['cmap'] = CommunityPalette(cmap[i]).colormap    
+
+    data.aes_dict['facecolor_kwargs']['c'] = 'white'
+    data.plot(ax=axs.flatten()[i],colorbar=True, outline=True, gridline=True, contourf=True)
