+++ src/data/colormaps/Speed.txt
@@ -1,0 +1,256 @@
+#7B8694
+#7C8795
+#7E8996
+#7F8A98
+#818C99
+#838E9B
+#848F9C
+#86919E
+#87929F
+#8994A1
+#8B96A2
+#8C97A4
+#8E99A5
+#8F9AA7
+#919CA8
+#939EAA
+#949FAB
+#96A1AD
+#97A2AE
+#99A4B0
+#9BA6B1
+#9CA7B3
+#9EA9B4
+#9FAAB6
+#A1ACB7
+#A3AEB9
+#A4AFBA
+#A5B1BB
+#A6B2BC
+#A8B3BD
+#A9B5BE
+#AAB6C0
+#ABB7C1
+#ACB9C2
+#AEBAC3
+#AFBBC4
+#B0BDC5
+#B1BEC6
+#B2BFC7
+#B3C1C8
+#B5C2C9
+#B6C3CB
+#B7C4CC
+#B8C6CD
+#B9C7CE
+#BAC9CF
+#BCCAD0
+#BDCBD1
+#BECCD2
+#BFCED3
+#C0CFD4
+#C2D1D6
+#C2D2D6
+#C3D3D7
+#C4D4D8
+#C5D5D8
+#C6D6D9
+#C7D7DA
+#C8D8DB
+#C8D9DB
+#C9DADC
+#CADBDD
+#CBDCDE
+#CCDDDE
+#CDDEDF
+#CEDFE0
+#CEE0E1
+#CFE1E1
+#D0E2E2
+#D1E3E3
+#D2E4E4
+#D3E5E4
+#D4E6E5
+#D4E7E6
+#D5E8E7
+#D6E9E7
+#D7EAE8
+#D8EBE9
+#DAECE9
+#DBECE9
+#DDEDE9
+#DEEEEA
+#E0EFEA
+#E1F0EA
+#E3F0EB
+#E4F1EB
+#E6F2EB
+#E8F3EB
+#E9F4EC
+#EBF4EC
+#ECF5EC
+#EEF6EC
+#EFF7ED
+#F1F7ED
+#F2F8ED
+#F4F9EE
+#F5FAEE
+#F7FBEE
+#F8FBEE
+#FAFCEF
+#FBFDEF
+#FDFEEF
+#FFFFF0
+#FEFEED
+#FEFDEB
+#FEFCE8
+#FEFBE6
+#FEFBE4
+#FEFAE1
+#FEF9DF
+#FEF8DC
+#FDF7DA
+#FDF7D8
+#FDF6D5
+#FDF5D3
+#FDF4D0
+#FDF4CE
+#FDF3CC
+#FDF2C9
+#FDF1C7
+#FCF0C4
+#FCF0C2
+#FCEFC0
+#FCEEBD
+#FCEDBB
+#FCECB8
+#FCECB6
+#FCEBB4
+#FBEAB1
+#FBE9AD
+#FBE8A9
+#FAE7A5
+#FAE6A1
+#F9E59E
+#F9E49A
+#F9E396
+#F8E292
+#F8E18E
+#F7E08B
+#F7DF87
+#F7DE83
+#F6DD7F
+#F6DC7B
+#F5DB78
+#F5DA74
+#F5D970
+#F4D86C
+#F4D768
+#F3D665
+#F3D561
+#F3D45D
+#F2D359
+#F2D255
+#F2D252
+#F1D051
+#F1CE50
+#F0CC4F
+#F0CA4E
+#EFC84D
+#EFC64C
+#EEC44B
+#EEC24A
+#EDC049
+#EDBE48
+#ECBC47
+#ECBA46
+#EBB845
+#EBB644
+#EAB443
+#EAB242
[…]
