### 2024-05-08 f3b8249
better plot for model-data comparison
