### 2024-05-08 b7dfc2d
better 1D plot for all related scripts
