+++ docs/source/auto_examples/plot_arrcomparison.py
@@ -3,17 +3,16 @@
 Plot Comparison Between Arrays
 =========================================
 
-This example shows how to use the ArrComparison class to compare two arrays and plot the comparison.
+This example shows how to build the ArrComparison class plot the comparison.
 """
 
 from cgeniepy.skill import ArrComparison
 import numpy as np
 
 np.random.seed(1923)
-x = np.random.rand(36, 36)
-y = x + np.random.rand(36, 36)
+x = np.random.rand(100)
+y = x + np.random.rand(100)
 
 ## calculate skill score
 ac = ArrComparison(x, y)
-print("Pearson coefficient:", ac.pearson_r())
-ac.plot()
+ac.plot(marker='x')
