### 2024-05-08 e75a35a
add mutable attribution to ScatterData
