### 2024-05-08 3a0850d
update the plot_scatter example
