### 2024-05-08 1169d28
update plot_additional_cmap.py doc
