### 2024-05-08 facf539
make add_zebra_frame independent from the Vis class
