### 2024-05-08 1964873
change default value of add_border_ticks
