### 2024-05-08 87ccbc5
add docstring for add_border_ticks
