### 2024-05-08 32e16b6
update plot_scatter_map example
