+++ docs/source/auto_examples/index.rst
@@ -31,23 +31,6 @@ Below is a gallery of examples
 
 .. raw:: html
 
-    <div class="sphx-glr-thumbcontainer" tooltip="This example shows how to plot the 2D ScatterData object. I use a LGM d13C data from Peterson e...">
-
-.. only:: html
-
-  .. image:: /auto_examples/images/thumb/sphx_glr_plot_scatter_map_thumb.png
-    :alt:
-
-  :ref:`sphx_glr_auto_examples_plot_scatter_map.py`
-
-.. raw:: html
-
-      <div class="sphx-glr-thumbnail-title">Plot the 2D ScatterData</div>
-    </div>
-
-
-.. raw:: html
-
     <div class="sphx-glr-thumbcontainer" tooltip="Plot the model variable for a spficic basin, here we use sea surface temperature in the Atlanti...">
 
 .. only:: html
@@ -60,6 +43,23 @@ Below is a gallery of examples
 .. raw:: html
 
       <div class="sphx-glr-thumbnail-title">Plot basin-specific data</div>
+    </div>
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="This example shows how to plot the 2D ScatterData object. I use a LGM d13C data from Peterson e...">
+
+.. only:: html
+
+  .. image:: /auto_examples/images/thumb/sphx_glr_plot_scatter_map_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_auto_examples_plot_scatter_map.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Plot the 2D ScatterData</div>
     </div>
 
 
@@ -276,8 +276,8 @@ Below is a gallery of examples
    :hidden:
 
    /auto_examples/plot_arrcomparison
-   /auto_examples/plot_scatter_map
    /auto_examples/plot_pac_sst
+   /auto_examples/plot_scatter_map
    /auto_examples/plot_coastal_data
    /auto_examples/plot_sst_map
    /auto_examples/plot_timeseries
