### 2024-05-08 4532c6a
fix a bug, not import matplotlib in example
