+++ examples/plot_scatter_map.py
@@ -8,6 +8,8 @@ This example shows how to plot the 2D ScatterData object. I use a LGM d13C data 
 
 from cgeniepy.table import ScatterData
 from cgeniepy.plot import CommunityPalette
+import matplotlib.pyplot as plt
+
 proxy_d13C = ScatterData("~/Science/lgm_bcp/data/LGM_d13c_CLEAN.txt", delimiter="\t", header=None)
 proxy_d13C.data.columns = ["Longitude", "Latitude", "Depth", "d13C","Event"]
 proxy_d13C.set_index(["Latitude", "Longitude"])
