### 2024-05-08 2a309a9
use lightgrey as default landmask facecolor
