### 2024-05-08 b30468f
add citation in about.rst
