### 2024-05-09 #6668 @rwegener2
👋🏻 @jimmielin & @koldunovn, would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
