### 2024-05-10 21a0785
add option to convert get_ts result to ScatData
