### 2024-05-10 bf54fdb
add warning: 'kind' argument in map/transect plot
