### 2024-05-17 #6762 @editorialbot
[REVIEW]: cgeniepy: A Python package for analysing cGENIE Earth System Model output
**Submitting author:** @ruiying-ocean (<a href="http://orcid.org/0000-0001-5811-2388">Rui Ying</a>)
**Repository:** https://github.com/ruiying-ocean/cgeniepy
**Branch with paper.md** (empty if default branch): master
**Version:** v0.14.2
