### 2024-07-05 #6762 @jimmielin
Hi @ruiying-ocean and @rwegener2,

Apologies for the delay. The software and the description manuscript looks good. I just have a few questions regarding the implementation / package design that maybe could be clarified in the manuscript.

* You mention there is a `Interpolator` class - could you clarify what it is used for and if you used any packages for this?
* It would be useful to provide some examples of model-data comparison, whether your package can ingest observational data for comparisons, example of the skill score calculation, etc.

Otherwise it looks good and I'm happy to accept with minor revisions. Great work!
Haipeng
