+++ publication/paper.md
@@ -18,28 +18,37 @@ affiliations:
  - name: School of Earth Sciences, University of Bristol, UK
    index: 1  
 
-date: 17 Feburary 2024
+date: 17 August 2024
 bibliography: paper.bib
 ---
 
 # Summary
 
-The cGENIE model is an numerical model that simulates the various components of the Earth system (e.g., atmosphere, ocean, land, biosphere, ice sheet) and their interactions [@ridgwell2007]. It has been widely used in studying and reconstructing the past ocean and climate states. Here, I provide a Python package *cgeniepy* for reading, analysing, and visualising the cGENIE model output, and performing the model-data comparison. The package is designed to facilitate the post-simulation analysis for all cGENIE users, as used in my recent studies [@ying2023;@ying2023b]. The package is designed as object-oriented, thus many features are standalone and can be used without cGENIE background.
+cGENIE Earth system model is an numerical model that simulates the various components of the Earth system (e.g., atmosphere, ocean, land, biosphere, ice sheet) and their interactions [@ridgwell2007]. It has been widely used in studying and reconstructing the past ocean and climate states. Here, I provide a Python package *cgeniepy* for reading, analysing, and visualising the cGENIE model output, and performing the model-data comparison. The package is designed to facilitate the post-simulation analysis for all cGENIE users, as used in my recent studies [@ying2023;@ying2023b]. The package is designed as object-oriented, thus many features are standalone and can be used without cGENIE background.
 
 # Statement of need
 Earth System Models (ESM) are the essential tool used to study the mechanism regulating the complex climate and their impacts. cGENIE is such a model with reduced model complexity and enhanced running speed that strengthen its application in paleoceanography studies. For instance, @henehan2019 used the model to study the impact of an extreme event (Cretaceous-Paleogene massive extinction in 66 Million years ago). @pohl2022 used it to study the long-term evolution of ocean oxygen in the past 550 million years. The application of this model has provided invaluable implication to study the evolution of climate and life.
 
-However, despite the power of cGENIE, the analysis of the model output has no consistent and systematic package support yet. Such gap could influence the efficiency and reproducibility of the research, in particular for users who are not familiar with the model output format or need to perform advanced analysis (e.g., model ensemble and model-skill assessment). Python is a open-source modern programming language that has built-in package management system, which makes it easy to install and use. The *cgeniepy* package provides a starting point for the continuously growing cGENIE community to develop and maintain a set of convenient tools, as seen in the other general circulation models (e.g., @romain2023 for the NEMO model and @gael2023 for the MITgcm model).
+Despite the power of cGENIE, the analysis of its model output has relied on a collection of Matlab scripts developed by the cGENIE maintainer (https://github.com/derpycode/muffinplot). A systematic package has been long missing. Such gap could influence the efficiency and reproducibility of the research, in particular for users who are not familiar with the Matlab language or need to perform advanced analysis (e.g., model ensemble based analysis). Python is a popular open-source modern programming language that has built-in package management system, which makes it easy to install, use, and advertise. The *cgeniepy* package provides a starting point for the continuously growing cGENIE community to develop and maintain a set of convenient tools, as seen in the other general circulation models (e.g., @romain2023 for the NEMO model and @gael2023 for the MITgcm model).
 
 # Package Design
[…]
