+++ publication/paper.bib
@@ -142,3 +142,21 @@
   keywords = {Carbon cycle,IMPORTANT,Palaeoceanography,Palaeoclimate,Read},
   file = {/Users/yingrui/Zotero/storage/7QRTU38F/Gutjahr et al. - 2017 - Very large release of mostly volcanic carbon durin.pdf}
 }
+
+@article{peterson2014,
+  title = {Deglacial Whole-Ocean {{δ13C}} Change Estimated from 480 Benthic Foraminiferal Records},
+  author = {Peterson, Carlye D. and Lisiecki, Lorraine E. and Stern, Joseph V.},
+  date = {2014},
+  journaltitle = {Paleoceanography},
+  volume = {29},
+  number = {6},
+  pages = {549--563},
+  issn = {1944-9186},
+  doi = {10.1002/2013PA002552},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/2013PA002552},
+  urldate = {2022-06-13},
+  abstract = {Terrestrial carbon storage is dramatically decreased during glacial periods due to cold temperatures, increased aridity, and the presence of large ice sheets on land. Most of the carbon released by the terrestrial biosphere is stored in the ocean, where the light isotopic signature of terrestrial carbon is observed as a 0.32–0.7‰ depletion in benthic foraminiferal δ13C. The wide range in estimated δ13C change results from the use of different subsets of benthic δ13C data and different methods of weighting the mean δ13C by volume. We present a more precise estimate of glacial-interglacial δ13C change of marine dissolved inorganic carbon using benthic Cibicidoides spp. δ13C records from 480 core sites (more than 3 times as many sites as previous studies). We divide the ocean into eight regions to generate linear regressions of regional δ13C versus depth for the Late Holocene (0–6 ka) and Last Glacial Maximum (19–23 ka) and estimate a mean δ13C decrease of 0.38 ± 0.08‰ (2σ) for 0.5–5 km. Estimating large uncertainty ranges for δ13C change in the top 0.5 km, below 5 km, and in the Southern Ocean, we calculate a whole-ocean change of 0.34 ± 0.19‰. This implies a terrestrial carbon change that is consistent with recent vegetation model estimates of 330–694 Gt C. Additionally, we find that a well-constrained surface ocean δ13C change is essential for narrowing the uncertainty range of estimated whole-ocean δ13C change.},
+  langid = {english},
+  keywords = {benthic,carbon isotope,carbon storage change,deglacial,Holocene,LGM},
+  file = {/Users/yingrui/Zotero/storage/QT5FQ646/Peterson et al. - 2014 - Deglacial whole-ocean δ13C change estimated from 4.pdf}
+}
