+++ .github/workflows/draft-pdf.yml
@@ -1,3 +1,4 @@
+name: Draft PDF
 on: [push]
 
 jobs:
@@ -12,12 +13,12 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: publication/paper.md
+          paper-path: paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: publication/paper.pdf
+          path: paper.pdf
