### 2024-08-17 fddf951
label the figures and add autoref
