### 2024-08-17 #6762 @ruiying-ocean
Hi @jimmielin,

Thanks for your appreciation and useful comments!

For your questions,
1. The `Interpolator` class is based on `scipy.interpolate` and it is designed for increasing the grid resolution of cGENIE and creating prettier figure (e.g., the vertical distribution of nutrient, O2, https://cgeniepy.readthedocs.io/en/latest/auto_examples/plot_PO4_distribution.html#sphx-glr-auto-examples-plot-po4-distribution-py)
2. I have added a section for this according to your suggestion.

Both are incorporated in the updated manuscript, which can be found in the https://github.com/ruiying-ocean/cgeniepy/actions if you go to the latest sucessful build.

> Hi @ruiying-ocean and @rwegener2,
>
> Apologies for the delay. The software and the description manuscript looks good. I just have a few questions regarding the implementation / package design that maybe could be clarified in the manuscript.
>
>     * You mention there is a `Interpolator` class - could you clarify what it is used for and if you used any packages for this?
>
>     * It would be useful to provide some examples of model-data comparison, whether your package can ingest observational data for comparisons, example of the skill score calculation, etc.
>
>
> Otherwise it looks good and I'm happy to accept with minor revisions. Great work! Haipeng
