### 2024-09-05 43540d4
fix:plot.py not working when one dimension (time) has only 1 value
