+++ examples/plot_PO4_distribution.py
@@ -23,7 +23,6 @@ from cgeniepy.plot import CommunityPalette
 import matplotlib.pyplot as plt
 
 model = GenieModel("/Users/yingrui/Science/lgm_foram_niche/model/muffin.CBE.worlg4.BASESFeTDTL.SPIN")
-ocn_po4 = model.get_var("ocn_PO4").isel(time=-1)
 
 fig, axs=plt.subplots(nrows=3, ncols=1, figsize=(6,9), tight_layout=True)
 
