### 2024-09-05 5a8ed88
remove an line of print used in debug
