### 2024-09-14 ruiying-ocean/cgeniepy#1 @koldunovn
Data examples
I can't run any of the examples, since data for those examples are not available. Can you, please, put sample data, that allow to run code in 'examples' somewhere people outside you organisation can download?

openjournals/joss-reviews#6762
