### 2024-09-14 ruiying-ocean/cgeniepy#2 @koldunovn
Notebook in tests have failing cells
This is notebook in question: https://github.com/ruiying-ocean/cgeniepy/blob/master/test/test.ipynb . It it probably also related to #1

openjournals/joss-reviews#6762
