### 2024-09-14 ruiying-ocean/cgeniepy#3 @koldunovn
Several tests are failing
I tried to run pytest on the code, and several tests are failing:

There is also no automatic tests running on PR (with GitHub Actions), I think it would be worth to add them. Failing tests most probably related to #1  - so ether having sample dataset as part of the repository, or provide link to where download sample data would be great.

openjournals/joss-reviews#6762
