### 2024-09-14 ruiying-ocean/cgeniepy#1 @ruiying-ocean
There are sample data in this directory: https://github.com/ruiying-ocean/cgeniepy/tree/master/src/data/muffin.CBE.worlg4.BASESFeTDTL.SPIN

or, you may download it from zenodo, which is more complete https://zenodo.org/records/10575295
