### 2024-09-14 ruiying-ocean/cgeniepy#1 @koldunovn
Thanks! Can you, please, update the download instructions, that it is clear where the data are and how to download them?
