### 2024-09-14 ruiying-ocean/cgeniepy#2 @koldunovn
Also `seaborn` is necessary for the notebook to work, but it is not in the requirements of the package.
