### 2024-09-14 ruiying-ocean/cgeniepy#2 @koldunovn
`LGM_d13c_CLEAN.txt` file references in the notebook is nowhere to be found.
