### 2024-09-14 ruiying-ocean/cgeniepy#2 @koldunovn
All paths in the notebook are absolute and point to author's directory. Please, at least define something like base path:
```
base_path='/Users/yingrui/Science/lgm_foram_niche/model/'
```
and use it for providing paths to the data.

Also, please, add instructions on how to download the data. I general, user should be able to just execute the cells and get the results, without too much additional editing of the notebook.
