### 2024-09-16 d0e7eee
remove previous src/data/muffin.CBE.worlg4.SPIN
