+++ src/data/muffin.CBE.worlg4.BASESFeTDTL.SPIN/ecogem/Plankton_params.txt
@@ -1,20 +1,0 @@
-PFT                 diameter      volume      vmax_C      qmin_P      qmax_P    vmax_PO4   affin_PO4      kexc_P     qmin_Fe     qmax_Fe     vmax_Fe    affin_Fe     kexc_Fe         q_C    alphachl  max_graz_C        kg_C      pp_opt      pp_sig respiration     biosink   mortality   beta_graz   beta_mort
-phytoplankton      0.600E+00   0.113E+00   0.225E+00   0.333E-02   0.110E-01   0.386E-01   0.236E+01   0.000E+00   0.100E-05   0.400E-05   0.170E-03   0.384E+00   0.000E+00   0.213E-11   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.798E+00   0.798E+00
-phytoplankton      0.190E+01   0.359E+01   0.114E+01   0.333E-02   0.110E-01   0.475E-01   0.703E+00   0.000E+00   0.100E-05   0.400E-05   0.125E-03   0.110E+00   0.000E+00   0.447E-10   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.793E+00   0.793E+00
-phytoplankton      0.600E+01   0.113E+03   0.363E+01   0.333E-02   0.110E-01   0.584E-01   0.210E+00   0.000E+00   0.100E-05   0.400E-05   0.915E-04   0.319E-01   0.000E+00   0.930E-09   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.777E+00   0.777E+00
-phytoplankton      0.190E+02   0.359E+04   0.161E+01   0.333E-02   0.110E-01   0.719E-01   0.627E-01   0.000E+00   0.100E-05   0.400E-05   0.670E-04   0.919E-02   0.000E+00   0.195E-07   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.736E+00   0.736E+00
-phytoplankton      0.600E+02   0.113E+06   0.718E+00   0.333E-02   0.110E-01   0.884E-01   0.187E-01   0.000E+00   0.100E-05   0.400E-05   0.491E-04   0.265E-02   0.000E+00   0.406E-06   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.650E+00   0.650E+00
-phytoplankton      0.190E+03   0.359E+07   0.418E+00   0.333E-02   0.110E-01   0.109E+00   0.559E-02   0.000E+00   0.100E-05   0.400E-05   0.360E-04   0.764E-03   0.000E+00   0.851E-05   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.538E+00   0.538E+00
-phytoplankton      0.600E+03   0.113E+09   0.284E+00   0.333E-02   0.110E-01   0.134E+00   0.167E-02   0.000E+00   0.100E-05   0.400E-05   0.264E-04   0.221E-03   0.000E+00   0.177E-03   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.457E+00   0.457E+00
[…]
