### 2024-09-16 9ffaaee
add sample_model
