+++ src/data/sample_model/ecogem/Plankton_params.txt
@@ -1,0 +1,12 @@
+PFT                 diameter      volume      vmax_C      qmin_P      qmax_P    vmax_PO4   affin_PO4      kexc_P     qmin_Fe     qmax_Fe     vmax_Fe    affin_Fe     kexc_Fe     qmin_Si     qmax_Si   vmax_SiO4  affin_SiO4     kexc_Si         q_C    alphachl  max_graz_C        kg_C      pp_opt      pp_sig respiration     biosink   mortality   beta_graz   beta_mort
+diatom             0.200E+01   0.419E+01   0.348E+01   0.333E-02   0.110E-01   0.479E-01   0.666E+00   0.000E+00   0.100E-05   0.400E-05   0.123E-03   0.104E+00   0.000E+00   0.336E-01   0.176E+00   0.479E-01   0.236E+01   0.000E+00   0.511E-10   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.792E+00   0.792E+00
+diatom             0.200E+02   0.419E+04   0.200E+01   0.333E-02   0.110E-01   0.726E-01   0.594E-01   0.000E+00   0.100E-05   0.400E-05   0.661E-04   0.869E-02   0.000E+00   0.336E-01   0.176E+00   0.726E-01   0.211E+00   0.000E+00   0.223E-07   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.733E+00   0.733E+00
+diatom             0.200E+03   0.419E+07   0.115E+01   0.333E-02   0.110E-01   0.110E+00   0.529E-02   0.000E+00   0.100E-05   0.400E-05   0.355E-04   0.723E-03   0.000E+00   0.336E-01   0.176E+00   0.110E+00   0.188E-01   0.000E+00   0.975E-05   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.533E+00   0.533E+00
+picoplankton       0.600E+00   0.113E+00   0.756E+00   0.333E-02   0.110E-01   0.386E-01   0.236E+01   0.000E+00   0.100E-05   0.400E-05   0.170E-03   0.384E+00   0.000E+00   0.000E+00   0.000E+00   0.000E+00   0.836E+01   0.000E+00   0.213E-11   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.798E+00   0.798E+00
+picoplankton       0.200E+01   0.419E+01   0.101E+01   0.333E-02   0.110E-01   0.479E-01   0.666E+00   0.000E+00   0.100E-05   0.400E-05   0.123E-03   0.104E+00   0.000E+00   0.000E+00   0.000E+00   0.000E+00   0.236E+01   0.000E+00   0.511E-10   0.383E-06   0.000E+00   0.500E+01   0.100E+02   0.200E+01   0.000E+00   0.000E+00   0.500E-01   0.792E+00   0.792E+00
[…]
