### 2024-09-17 ruiying-ocean/cgeniepy#1 @ruiying-ocean
I have written a function to load the sample model, also updated in the readme

```python
import cgeniepy

model = cgeniepy.sample_model()
model.get_var('ocn_sur_temp').isel(time=-1).plot(colorbar=True)
```
