### 2024-09-17 ruiying-ocean/cgeniepy#2 @ruiying-ocean
This file is not meant to be uploaded. I use it to do some quick tests. Seaborn is not necessary for the package function per se.
So I will just remove this file and update the download instructions as #1.
