### 2024-09-17 901d688
resolve #2, remove test.ipynb
