### 2024-09-17 d1666ce
model download instruction corresponding to #1/#3
