+++ README.md
@@ -39,7 +39,7 @@ model = cgeniepy.sample_model()
 model.get_var('ocn_sur_temp').isel(time=-1).plot(colorbar=True)
 ```
 
-If you want to try other model runs, I have uploaded some in zenodo (https://zenodo.org/records/10575295). However, not all cGENIE model experiments are publicly available because many different research teams use cGENIE. Therefore, this is not what this package can resolve.
+If you want to try other model runs, I have uploaded some to zenodo (https://zenodo.org/records/10575295). However, not all cGENIE model experiments are publicly available because many different research teams use cGENIE. Therefore, this is not what this package can resolve.
 
 
 ## Full documentation
