### 2024-09-18 f625bff
resolve #3, update pytests
