+++ test/test_model.py
@@ -1,136 +1,17 @@
-from cgeniepy.model import GenieModel
-from importlib.resources import files
+import cgeniepy
 
 def test_model_getvar():
-    model_path = str(files("data").joinpath("muffin.CBE.worlg4.BASESFeTDTL.SPIN"))
-    model = GenieModel(model_path, gemflag=['ecogem'])
-    data= model.get_var('eco2D_Plankton_C_Total').isel(time=-1).mean().data.item()
-    assert data  == 1.690475344657898
+    model = cgeniepy.sample_model()
+    assert model.get_var("ocn_sur_temp").mean().data.values.item() == 18.055744171142578
+    
+def test_eco_pft():
+    model = cgeniepy.sample_model(model_type='EcoModel',gemflag=['ecogem'])
+    ## 1st PFT
+    data = model.get_pft(1).isel(time=-1).mean().data.values.item()
+    assert data  == 0.12259095907211304
 
-def tets_model_ncvardict():
-    model_path = str(files("data").joinpath("muffin.CBE.worlg4.BASESFeTDTL.SPIN"))
-    model = GenieModel(model_path, gemflag=['ecogem'])
-
-    data = model._ncvar_dict
-    baseline = {'./muffin.CBE.worlg4.BASESFeTDTL.SPIN/ecogem/fields_ecogem_2d.nc': ['time',
-    'year',
-    'lon',
-    'lat',
-    'zt',
-    'lon_edges',
-    'lat_edges',
-    'zt_edges',
-    'grid_level',
-    'grid_mask',
-    'grid_topo',
-    'eco2D_Plankton_C_001',
-    'eco2D_Plankton_C_002',
-    'eco2D_Plankton_C_003',
-    'eco2D_Plankton_C_004',
-    'eco2D_Plankton_C_005',
-    'eco2D_Plankton_C_006',
-    'eco2D_Plankton_C_007',
-    'eco2D_Plankton_C_008',
-    'eco2D_Plankton_C_009',
-    'eco2D_Plankton_C_010',
-    'eco2D_Plankton_C_011',
-    'eco2D_Plankton_C_012',
-    'eco2D_Plankton_C_013',
-    'eco2D_Plankton_C_014',
-    'eco2D_Plankton_C_015',
-    'eco2D_Plankton_C_016',
-    'eco2D_Plankton_C_017',
-    'eco2D_Plankton_C_018',
-    'eco2D_Plankton_C_019',
-    'eco2D_Plankton_C_Total',
-    'eco2D_Uptake_Fluxes_C',
-    'eco2D_Plankton_P_001',
-    'eco2D_xGamma_P_001',
-    'eco2D_Plankton_P_002',
-    'eco2D_xGamma_P_002',
-    'eco2D_Plankton_P_003',
-    'eco2D_xGamma_P_003',
-    'eco2D_Plankton_P_004',
-    'eco2D_xGamma_P_004',
-    'eco2D_Plankton_P_005',
-    'eco2D_xGamma_P_005',
-    'eco2D_Plankton_P_006',
-    'eco2D_xGamma_P_006',
-    'eco2D_Plankton_P_007',
-    'eco2D_xGamma_P_007',
-    'eco2D_Plankton_P_008',
-    'eco2D_xGamma_P_008',
-    'eco2D_Plankton_P_009',
-    'eco2D_Plankton_P_010',
-    'eco2D_Plankton_P_011',
-    'eco2D_Plankton_P_012',
[…]
