+++ test/test_skill.py
@@ -8,14 +8,14 @@ def create_testdata():
     ## calculate skill score
     return ArrComparison(x, y)    
 
-
 def test_mscore():    
     ac = create_testdata()
     assert ac.mscore()==1.0
 
 def test_pearson_r():
     ac = create_testdata()
-    assert ac.pearson_r()==0.9999999999999999
+    diff = ac.pearson_r().item() - 1.0
+    assert diff < 1E-8
 
 def test_cos_sim():    
     ac = create_testdata()
