### 2024-09-18 ruiying-ocean/cgeniepy#3 @ruiying-ocean
The package test is updated and a local result is as below
