### 2024-09-18 486e1b4
doc: update examples using cgeniepy.sample_model()
