+++ examples/plot_coastal_data.py
@@ -7,18 +7,18 @@ This example shows how to plot the PO4 distribution in each basin.
 """
 
 import xarray as xr
-from cgeniepy.model import GenieModel
+import cgeniepy
 import numpy as np
 
-pi_model = GenieModel("/Users/yingrui/Science/lgm_bcp/model/muffin.CB.worlg4.BASESFeTDTL.SPIN", gemflag='biogem')
+model = cgeniepy.sample_model()
 
-sst = pi_model.get_var('ocn_sur_temp').isel(time=-1)
+sst = model.get_var('ocn_sur_temp').isel(time=-1)
 
 ## grid category is a pre-defined 2D data
 ## 0: coastal
 ## 1: land
 ## 2: open ocean
-gc = pi_model.grid_category()
+gc = model.grid_category()
 
 ## coastal region
 xr.where(gc == 0, sst.data.values, np.nan).plot()
