+++ examples/plot_multiple_vars.py
@@ -26,7 +26,7 @@ fig, axs = plt.subplots(2, 2, figsize=(10, 8), subplot_kw={'projection': ccrs.Pl
 
 pi_model = cgeniepy.sample_model()
 
-variable = ['ocn_sur_temp', 'bio_export_POC', 'ocn_sur_PO4','ocn_sur_O2']
+variable = ['ocn_sur_temp', 'bio_fexport_POC', 'ocn_sur_PO4','ocn_sur_O2']
 cmap = ['ocean_temp', 'tol_rainbow', 'WtBuGnYlRd','kovesi_rainbow']
 
 for i in range(4):
