### 2024-09-18 1dce615
add more example data (related to  #1)
