+++ docs/source/auto_examples/plot_timeseries.py
@@ -3,7 +3,7 @@
 Extract and Plot cGENIE Time Series data
 ==============================================
 
-This example shows how to read in and plot cGENIE time series data
+This example shows how to read in and plot cGENIE time series data. The example historical run (1750-2022) is available from https://zenodo.org/records/10575295.
 """
 from cgeniepy.model import GenieModel
 import seaborn as sns
@@ -19,5 +19,5 @@ o2 = model.get_ts("ocn_O2")
 ## merge both and convert to ScatterData format
 ts_data = ScatterData(temp.merge(o2, on="time (yr)"))
 ts_data.set_index("time (yr)")
-ts_data.plot(var="temperature (C)", ax=axs[0])
-ts_data.plot(var="surface O2 (mol kg-1)", ax=axs[1])
+ts_data.plot(var="temperature (C)", ax=axs[0], kind='line')
+ts_data.plot(var="surface O2 (mol kg-1)", ax=axs[1], kind='line')
