### 2024-09-18 ruiying-ocean/cgeniepy#1 @koldunovn
Thank you! Can you, please, update code in all examples as well to use sample model, as some of them still use hard coded paths (e.g. `examples/plot_d13C_comp.py`, `examples/plot_timeseries.py`)
