+++ pyproject.toml
@@ -4,6 +4,9 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "cgeniepy"
-version = "0.13.3"
+version = "0.13.5"
 description = "A Python package to read, analyse and visualise cGENIE Earth System Model output"
 license = {text = "GPL-3.0-only"}
+
+[project.optional-dependencies]
+extra = ["pangaeapy"]
