+++ setup.cfg
@@ -1,8 +1,8 @@
 [metadata]
 name = cgeniepy
-version = 0.13.3
+version = 0.13.5
 author = Rui Ying
-author_email = rui.ying@bristol.ac.uk
+author_email = ying.rui@outlook.com
 description = A Python package to read, analyse and visualise cGENIE Earth System Model output
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -33,3 +33,7 @@ where = src
 
 [options.package_data]
 * = *.nc, *.csv, *.txt, *.xml
+
+[options.extras_require]
+extra = pangaeapy
+
