### 2024-09-19 7e2c0e5
fix #5, install extra dependency
