### 2024-09-19 1ed031f
update example to remove hard codes
