+++ test/test_table.py
@@ -18,3 +18,9 @@ def test_detectbasin():
     data.set_index(['lat','lon'])
     basin_value = data.detect_basin()['basin'].values.item()
     assert basin_value =='Indian Ocean'
+
+def test_url():
+    url = "https://www.ncei.noaa.gov/pub/data/paleo/icecore/antarctica/epica_domec/edc-co2-2008-bern-noaa.txt"
+    
+    test_data= ScatterData(url, comment='#', delimiter='\t')
+    assert test_data.data['CO2'][0] == 257.8
