### 2024-09-19 6462fbb
update plot_scatter_map.py to remove hard code
