+++ examples/plot_scatter_map.py
@@ -9,8 +9,13 @@ This example shows how to plot the 2D ScatterData object. I use a LGM d13C data 
 from cgeniepy.table import ScatterData
 from cgeniepy.plot import CommunityPalette
 import matplotlib.pyplot as plt
+import subprocess
 
-proxy_d13C = ScatterData("~/Science/lgm_bcp/data/LGM_d13c_CLEAN.txt", delimiter="\t", header=None)
+## Download necessary files, you need to install zenodo_get first
+## by `pip install zenodo_get`, or, just download it from the link above
+subprocess.call(["zenodo_get", "10.5281/zenodo.13786013", "-o", "~/Downloads/"])
+
+proxy_d13C = ScatterData("~/Downloads/LGM_d13c_CLEAN.txt", delimiter="\t", header=None)
 proxy_d13C.data.columns = ["Longitude", "Latitude", "Depth", "d13C","Event"]
 proxy_d13C.set_index(["Latitude", "Longitude"])
 cmap = CommunityPalette("BuDaRd18").colormap
