+++ docs/source/auto_examples/index.rst
@@ -83,23 +83,6 @@ Below is a gallery of examples
 
 .. raw:: html
 
-    <div class="sphx-glr-thumbcontainer" tooltip="This example shows how to plot the 2D ScatterData object. I use a LGM d13C data from Peterson et al. 2014 (https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2013PA002552) as an example.">
-
-.. only:: html
-
-  .. image:: /auto_examples/images/thumb/sphx_glr_plot_scatter_map_thumb.png
-    :alt:
-
-  :ref:`sphx_glr_auto_examples_plot_scatter_map.py`
-
-.. raw:: html
-
-      <div class="sphx-glr-thumbnail-title">Plot the 2D ScatterData</div>
-    </div>
-
-
-.. raw:: html
-
     <div class="sphx-glr-thumbcontainer" tooltip="This example shows how to read in and plot cGENIE time series data.">
 
 .. only:: html
@@ -129,6 +112,23 @@ Below is a gallery of examples
 .. raw:: html
 
       <div class="sphx-glr-thumbnail-title">Customise the 2D map projection</div>
+    </div>
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="This example shows how to plot the 2D ScatterData object. I use a LGM d13C data from Peterson et al. 2014 (https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2013PA002552) as an example.">
+
+.. only:: html
+
+  .. image:: /auto_examples/images/thumb/sphx_glr_plot_scatter_map_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_auto_examples_plot_scatter_map.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Plot the 2D ScatterData</div>
     </div>
 
 
@@ -282,9 +282,9 @@ Below is a gallery of examples
    /auto_examples/plot_pac_sst
    /auto_examples/plot_coastal_data
    /auto_examples/plot_sst_map
-   /auto_examples/plot_scatter_map
    /auto_examples/plot_timeseries
    /auto_examples/plot_logo
+   /auto_examples/plot_scatter_map
    /auto_examples/plot_scatterdata
    /auto_examples/plot_PO4_distribution
    /auto_examples/plot_multiple_vars
