### 2024-09-19 ruiying-ocean/cgeniepy#1 @ruiying-ocean
No problem! the hard codes are removed. Now the example files directly download the example dataset from my provided url. A new feature has been included to resolve this #5
