### 2024-09-19 #6762 @ruiying-ocean
> I have only one remaining issue on updating the examples [ruiying-ocean/cgeniepy#1](https://github.com/ruiying-ocean/cgeniepy/issues/1) As soon as it is addressed we are good to go.

The example scripts are updated to avoid hard codes. Please check them and I appreciate your help!
