+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = cgeniepy
-version = 0.13.5
+version = 0.13.6
 author = Rui Ying
 author_email = ying.rui@outlook.com
 description = A Python package to read, analyse and visualise cGENIE Earth System Model output
