### 2024-09-19 #6762 @ruiying-ocean
@rwegener2 Rachel do you know how long it takes to get a doi after this paper being accepted? I am now pre-proofing another paper and I am happy to update the reference if this is within 2 weeks.
