### 2024-09-19 #6762 @rwegener2
> Rachel do you know how long it takes to get a doi after this paper being accepted? I am now pre-proofing another paper and I am happy to update the reference if this is within 2 weeks.

@ruiying-ocean If reviewers are satisfied with submission it usually takes 3-7 days to finalize the publication/get a DOI. After reviewers confirm approval I generate a final checklist that will include items for you and for myself. I'll ask you to upload your code to Zenodo (or similar) and generate a code DOI at that point. Once I you generate DOI and complete my checks the submission gets sent back to the Track Editor, who finalizes the paper publication and closes the issue. The paper will get a DOI when it is published.
