### 2024-09-19 #6762 @rwegener2
Editor checks paper proof:
- [ ] Download the proof, check all references have DOIs, follow the links and check the references

Editor checks archive generated by author:
- [ ] Check that the archive title ✅, author list ✅, version tag ✅, and the license ✅ are correct

Editor double checks paper and recommends submission:
