### 2024-09-19 5e5dfd6
remove hard code in paper.md, fix some typos
