### 2024-09-19 #6762 @ruiying-ocean
@rwegener2
* The tagged release is created as v0.14.0 (With this release I have modified some typos in my paper. So it can be slightly different from the generated one by editorbot )
* The code is archived in Zenodo, with DOI: https://doi.org/10.5281/zenodo.13799839
* The title/author/affiliations/ORCID/License is confirmed.

Let me know if anything is needed, thank you!
