### 2024-09-19 4774c39
Fix small typos in paper.md
