### 2024-09-19 a342b42
Fix additional small typos in paper.md
