### 2024-09-19 #6762 @rwegener2
Hi @ruiying-ocean 👋🏻 I read through the manuscript. It looks good! I just made a few small changes and comments on confusing lines in https://github.com/ruiying-ocean/cgeniepy/pull/7. Let me know if you have anything isn't clear.
