+++ publication/paper.bib
@@ -45,6 +45,7 @@
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
+  doi = {10.5281/zenodo.5724577},
   url={https://github.com/rcaneill/xnemogcm}
 }
 
@@ -54,6 +55,7 @@
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
+  doi = {10.5281/zenodo.13766391},
   url={https://github.com/gaelforget/MITgcmTools.jl}
 }
 
