### 2024-09-24 5e1d5db
add quarto file and fix a typo
