### 2024-09-25 #6762 @rwegener2
> I have published a new release and it will appear soon in the same zenodo record

Thanks @ruiying-ocean! So the correct version tag is now v0.14.2, correct? And the DOI is https://zenodo.org/records/13835532?
