### 2024-09-26 #6762 @ruiying-ocean
> > I have published a new release and it will appear soon in the same zenodo record
>
> Thanks @ruiying-ocean! So the correct version tag is now v0.14.2, correct? And the DOI is https://doi.org/10.5281/zenodo.13835532?

Yes! THat is correct
