### 2024-09-26 #6762 @rwegener2
Hi @ruiying-ocean, looks like there are a few DOI errors.

It looks like the 3 citations listed are [Handles](https://www.doi.org/the-identifier/resources/factsheets/doi-system-and-the-handle-system), but not DOIs. I checked the first two and it looks like there are DOIs available for those resources:
- `https://doi.org/10.1073/pnas.1905989116` instead of `10/ggbnrm`
- `https://doi.org/10.5194/bg-4-87-2007` instead of `10/dcq62d`

Would you please look for the DOI for the last citation and replace these three citations in your bibtex file with the `10.xxxx/` format instead?
