### 2024-09-26 b9bcb71
correct doi in bibtex file
