### 2024-09-26 #6762 @rwegener2
@ruiying-ocean I'm sorry I didn't say more specifically - Please put just the part after doi.org gets put in the bibtex citation (ex. `10.1073/pnas.1905989116`)
