### 2024-09-27 1d073cc
doi pattern 10.xxx
