### 2024-09-27 #6762 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
