### 2024-09-30 #6762 @kthyng
@ruiying-ocean:
* Can you update the Zenodo metadata so the author list doesn't include your JOSS editor? Also usually your title would just be "cgeniepy: A Python package for analysing cGENIE Earth System Model output" without the "ruiying-ocean/cgeniepy:" beforehand, but that is up to you.
