### 2024-09-30 eed5eb2
Update paper.md
Grammar corrections
