### 2024-09-30 #6762 @ruiying-ocean
Hi @kthyng The pull request is merged. Do I need to create a new release to zenodo?
