### 2024-09-30 #6762 @kthyng
@ruiying-ocean Paper looks good! I'll wait to hear back about the Zenodo metadata.
