### 2024-09-30 #6762 @ruiying-ocean
@kthyng  Great! the metadata in Zenodo was just updated
