### 2024-09-30 #6762 @ruiying-ocean
How exciting, thank you and thanks everyone for helping!

> Congratulations on your new publication @ruiying-ocean! Many thanks to editor @rwegener2 and to reviewers @koldunovn and @jimmielin for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
>
> @ruiying-ocean If you'd like to join JOSS as a reviewer, please sign up https://reviewers.joss.theoj.org/!
